rename files cuda -> gpu
