+++ arbor/backends/gpu/gpu_common.hpp
