+++ arbor/backends/gpu/matrix_interleave.hpp
@@ -4,7 +4,7 @@
 #include <hip/hip_runtime.h>
 #endif
 
-#include "cuda_common.hpp"
+#include "gpu_common.hpp"
 #include "matrix_common.hpp"
 
 namespace arb {
