+++ arbor/backends/gpu/matrix_solve.cu
@@ -4,7 +4,7 @@
 
 #include <arbor/fvm_types.hpp>
 
-#include "cuda_common.hpp"
+#include "gpu_common.hpp"
 #include "matrix_common.hpp"
 
 namespace arb {
