+++ arbor/backends/gpu/mechanism.cu
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <backends/event.hpp>
 #include <backends/multi_event_stream_state.hpp>
-#include <backends/gpu/cuda_common.hpp>
+#include <backends/gpu/gpu_common.hpp>
 #include <backends/gpu/math_cu.hpp>
 #include <backends/gpu/mechanism_ppack_base.hpp>
 #include <backends/gpu/reduce_by_key.hpp>
