+++ arbor/backends/gpu/reduce_by_key.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <cstdint>
-#include "cuda_atomic.hpp"
-#include "cuda_common.hpp"
+#include "gpu_atomic.hpp"
+#include "gpu_common.hpp"
 
 namespace arb {
 namespace gpu {
