+++ arbor/backends/gpu/shared_state.cu
@@ -8,8 +8,8 @@
 #include <backends/event.hpp>
 #include <backends/multi_event_stream_state.hpp>
 
-#include "cuda_atomic.hpp"
-#include "cuda_common.hpp"
+#include "gpu_atomic.hpp"
+#include "gpu_common.hpp"
 
 namespace arb {
 namespace gpu {
